Bulletin Board overall look
First commit but worked on this the entire weekend. Found some css and html styling to create the corkboard look as well as sticky note look. Went in and updated it to fit our needs. Toggle blocked items/pages that didn't seem to be needed as I didin't want to get rid of anything just in case we might need it in the future.